refactor: UserService 인터페이스 추가 및 구현체 추가
